// apps/client/src/constants/tracking.ts
/**
 * Constants for tracking and analytics events
 */

export enum UpsellEventSubCategory {
    IMMEDIATE_PRE_ROLL = "immediate pre roll",
    HUB_PRE_ROLL = "hub pre roll",
}

export enum MobileHubEventName {
    WEB_APP_SCREEN_DISPLAYED = "Web App Screen Displayed",
    WEB_APP_BUTTON_PRESSED = "Web App Button Pressed",
    WEB_APP_ERROR_OCCURRED = "Web App Error Occurred",
}